Xinyu database

+++ server/src/database.js
@@ -42,8 +42,8 @@ function insertUserId(user_id, callback) {
 // delete user_id from user table
 // should able to delete all data related to the user needs to be deleted
 function deleteUserId(user_id, callback) {
-  let sql = `DELETE FROM user WHERE ?? = ?`;
-  let inserts = ["user_id", user_id];
+  let sql = `DELETE FROM user WHERE user_id = ?`;
+  let inserts = [user_id];
   connection.query(sql, inserts, function (err, result) {
     // check error type later
     if (err) {
@@ -63,8 +63,8 @@ function deleteUserId(user_id, callback) {
 
 // return list of JSON contains all tenant info for a property
 function selectAllTenants(property_id, callback) {
-  let sql = `SELECT * FROM tenant WHERE ?? = ?`;
-  let inserts = ["property_id", property_id];
+  let sql = `SELECT * FROM tenant WHERE property_id = ?`;
+  let inserts = [property_id];
   connection.query(sql, inserts, function (err, tenantList) {
     if (err) {
       console.log(`not able to select tenantList of property_id: ${property_id} from database`);
@@ -87,13 +87,8 @@ function updateTenant(tenant_id, update_info, callback) {
   let landlord_phone = update_info.landlord_phone;
   let rubs = update_info.rubs;
 
-  let sql = `UPDATE tenant set ?? = ?, ?? = ?, ?? = ?, ?? = ?, ?? = ? WHERE ?? = ?`;
-  let inserts = ["name", name,
-                 "email", email,
-                 "address", address,
-                 "landlord_phone", landlord_phone,
-                 "rubs", rubs,
-                 "tenant_id", tenant_id];
+  let sql = `UPDATE tenant set name = ?, email = ?, address = ?, landlord_phone = ?, rubs = ? WHERE tenant_id = ?`;
+  let inserts = [name, email, address, landlord_phone, rubs, tenant_id];
 
   connection.query(sql, inserts, function (err, result) {
     if (err) {
@@ -146,11 +141,9 @@ function insertTenant(property_id, tenant_info, callback) {
 // return true if delete successfully
 // return false if delete failed
 function deleteTenant(property_id, tenant_id, callback) {
-  let sql = `DELETE FROM tenant WHERE ?? = ? AND ?? = ?`;
-  let inserts = ["property_id", property_id,
-                 "tenant_id", tenant_id];
+  let sql = `DELETE FROM tenant WHERE property_id = ? AND tenant_id = ?`;
+  let inserts = [property_id, tenant_id];
   connection.query(sql, inserts, function (err, result) {
-    // check error type later
     if (err) {
       console.log(err);
       console.log(`not able to delete property_id: ${property_id} tenant_id: ${tenant_id} from database`);
@@ -169,8 +162,8 @@ function deleteTenant(property_id, tenant_id, callback) {
 
 // return a list of JSON contains all of the properties owned by user
 function selectAllProperties(user_id, callback) {
-  let sql = `SELECT * FROM property WHERE ?? = ?`;
-  let inserts = ["user_id", user_id];
+  let sql = `SELECT * FROM property WHERE user_id = ?`;
+  let inserts = [user_id];
   connection.query(sql, inserts, function (err, propertyList) {
     if (err) {
       console.log(`not able to select property of user_id: ${user_id} from database`);
@@ -225,14 +218,8 @@ function updateProperty(user_id, property_info, callback) {
   let landlord_phone = property_info.landlord_phone;
   let property_id = property_info.property_id;
 
-  let sql = `UPDATE property set ?? = ?, ?? = ?, ?? = ?, ?? = ? , ?? = ? WHERE ?? = ? AND ?? = ?`;
-  let inserts = ["name",  name,
-                 "address", address,
-                 "property_type", property_type,
-                 "total_footage", total_footage,
-                 "landlord_phone", landlord_phone,
-                 "user_id", user_id,
-                 "property_id", property_id];
+  let sql = `UPDATE property set name = ?, address = ?, property_type = ?, total_footage = ? , landlord_phone = ? WHERE user_id = ? AND property_id = ?`;
+  let inserts = [name, address, property_type, total_footage, landlord_phone, user_id, property_id];
 
   connection.query(sql, inserts, function (err, result) {
     if (err) {
@@ -254,11 +241,9 @@ function updateProperty(user_id, property_info, callback) {
 // return true if delete successfully
 // return false if delete fails
 function deleteProperty(user_id, property_id, callback) {
-  let sql = `DELETE FROM property WHERE ?? = ? AND ?? = ?`;
-  let inserts = ["property_id", property_id,
-                 "user_id", user_id];
+  let sql = `DELETE FROM property WHERE property_id = ? AND user_id = ?`;
+  let inserts = [property_id, user_id];
   connection.query(sql, inserts, function (err, result) {
-    // check error type later
     if (err) {
       console.log(err);
       console.log(`not able to delete property_id: ${property_id} user_id: ${user_id} from database`);
@@ -300,9 +285,8 @@ function insertMeter(property_id, meter_id, callback) {
 // return true if adds successfully
 // return false if adds failed
 function deleteMeter(property_id, meter_id, callback) {
-  let sql = `DELETE FROM meter WHERE ?? = ? AND ?? = ?`;
-  let inserts = ["property_id", property_id,
-                 "meter_id", meter_id];
+  let sql = `DELETE FROM meter WHERE property_id = ? AND meter_id = ?`;
+  let inserts = [property_id, meter_id];
   connection.query(sql, inserts, function (err, result) {
     if (err) {
       console.log(err);
@@ -321,8 +305,8 @@ function deleteMeter(property_id, meter_id, callback) {
 
 // return a list of JSON contains meter list of a property
 function selectAllMeters(property_id, callback) {
-  let sql = `SELECT meter_id FROM meter WHERE ?? = ?`;
-  let inserts = ["property_id", property_id];
+  let sql = `SELECT meter_id FROM meter WHERE property_id = ?`;
+  let inserts = [property_id];
   connection.query(sql, inserts, function (err, meterList) {
     if (err) {
       console.log(`not able to select meterList of property_id: ${property_id} from database`);
@@ -371,11 +355,8 @@ function updateSubmeter(submeter_id, update_info, callback){
   let meter_id = update_info.meter_id;
   let multiplier = update_info.multiplier;
 
-  let sql = `UPDATE tenant SET ?? = ?, ?? = ?, ?? = ? WHERE ?? = ?`;
-  let inserts = ["tenant_id", tenant_id,
-                 "meter_id", meter_id,
-                 "multiplier", multiplier,
-                 "submeter_id", submeter_id];
+  let sql = `UPDATE tenant SET tenant_id = ?, meter_id = ?, multiplier = ? WHERE submeter_id = ?`;
+  let inserts = [tenant_id, meter_id, multiplier, submeter_id];
   connection.query(sql, inserts, function (err, result) {
     if (err) {
       console.log(`not able to update submeter_id: ${submeter_id} into database`);
@@ -396,9 +377,8 @@ function updateSubmeter(submeter_id, update_info, callback){
 // return true if deletes successfully
 // return false if deletes failed
 function deleteSubmeter(tenant_id, submeter_id, callback) {
-  let sql = `DELETE FROM submeter WHERE ?? = ? AND ?? = ?`;
-  let inserts = ["tenant_id", tenant_id,
-                 "submeter_id", submeter_id];
+  let sql = `DELETE FROM submeter WHERE tenant_id = ? AND submeter_id = ?`;
+  let inserts = [tenant_id, submeter_id];
   connection.query(sql, inserts, function (err, result) {
     if (err) {
       console.log(err);
@@ -417,8 +397,8 @@ function deleteSubmeter(tenant_id, submeter_id, callback) {
 
 // return a list of JSON contains Submeter list of a tenant
 function selectAllSubmeters(tenant_id, callback) {
-  let sql = `SELECT submeter_id FROM submeter WHERE ?? = ?`;
-  let inserts = ["tenant_id", tenant_id];
+  let sql = `SELECT submeter_id FROM submeter WHERE tenant_id = ?`;
+  let inserts = [tenant_id];
   connection.query(sql, inserts, function (err, submeterList) {
     if (err) {
       console.log(`not able to select submeterList of tenant_id: ${tenant_id} from database`);
@@ -489,19 +469,8 @@ function updateBill(bill_id,bill_info, callback) {
   let total_charge = bill_info.total_charge;
   let unit_charge = bill_info.unit_charge;
 
-  let sql = `UPDATE bill set ?? = ?, ?? = ?, ?? = ?, ?? = ?, ?? = ?,?? = ?,?? = ?,?? = ?,?? = ?,?? = ?,?? = ? WHERE ?? = ?`;
-  let inserts = ["account_id", account_id,
-                 "meter_id", meter_id,
-                 "m_kwh_usage", m_kwh_usage,
-                 "from_date", from_date,
-                 "to_date", to_date,
-                 "m_charge", m_charge,
-                 "s_kwh_usage", s_kwh_usage,
-                 "s_charge", s_charge,
-                 "total_kwh_usage", total_kwh_usage,
-                 "total_charge", total_charge,
-                 "unit_charge", unit_charge,
-                 "bill_id", bill_id];
+  let sql = `UPDATE bill set account_id = ?, meter_id = ?, m_kwh_usage = ?, from_date = ?, to_date = ?,m_charge = ?,s_kwh_usage = ?,s_charge = ?,total_kwh_usage = ?,total_charge = ?,unit_charge = ? WHERE bill_id = ?`;
+  let inserts = [account_id, meter_id, m_kwh_usage, from_date, to_date, m_charge, s_kwh_usage, s_charge, total_kwh_usage, total_charge, unit_charge, bill_id];
   connection.query(sql, inserts, function (err, result) {
     if (err) {
       console.log(`not able to update bill info for bill_id: ${bill_id}  into database`);
@@ -522,8 +491,8 @@ function updateBill(bill_id,bill_info, callback) {
 // return true if delete successfully
 // return false if delete failed
 function deleteBill(bill_id, callback){
-  let sql = `DELETE FROM bill WHERE ?? = ?`;
-  let inserts = ["bill_id", bill_id];
+  let sql = `DELETE FROM bill WHERE bill_id = ?`;
+  let inserts = [bill_id];
   connection.query(sql, inserts, function (err, result) {
     if (err) {
       console.log(err);
@@ -591,17 +560,8 @@ function updateSubmeterBill(submeter_bill_id, updated_info, callback) {
   let amt_with_multiplier = updated_info.amt_with_multiplier;
   let amt_due = updated_info.amt_due;
 
-  let sql = `UPDATE submeter_bill SET ?? = ?, ?? = ?, ?? = ?, ?? = ?,?? = ?,?? = ?,?? = ?,?? = ?,?? = ? WHERE ?? = ?`;
-  let inserts = ["bill_id", bill_id,
-                 "submeter_id", submeter_id,
-                 "prior_read", prior_read,
-                 "cur_read", cur_read,
-                 "from_date", from_date,
-                 "to_date", to_date,
-                 "cur_amt", cur_amt,
-                 "amt_with_multiplier", amt_with_multiplier,
-                 "amt_due", amt_due,
-                 "submeter_bill_id", submeter_bill_id];
+  let sql = `UPDATE submeter_bill SET bill_id = ?, submeter_id = ?, prior_read = ?, cur_read = ?,from_date = ?,to_date = ?,cur_amt = ?,amt_with_multiplier = ?,amt_due = ? WHERE submeter_bill_id = ?`;
+  let inserts = [bill_id, submeter_id, prior_read, cur_read, from_date, to_date, cur_amt, amt_with_multiplier, amt_due, submeter_bill_id];
   connection.query(sql, inserts, function (err, result) {
     if (err) {
       console.log(`not able to update submeter_bill info for submeter_bill_id: ${submeter_bill_id} into database`);
@@ -622,8 +582,8 @@ function updateSubmeterBill(submeter_bill_id, updated_info, callback) {
 // return true if delete successfully
 // return false if delete failed
 function deleteSubmeterBill(submeter_bill_id, callback){
-  let sql = `DELETE FROM submeter_bill WHERE ?? = ?`;
-  let inserts = ["submeter_bill_id", submeter_bill_id];
+  let sql = `DELETE FROM submeter_bill WHERE submeter_bill_id = ?`;
+  let inserts = [submeter_bill_id];
   connection.query(sql, inserts, function (err, result) {
     if (err) {
       console.log(err);
@@ -693,18 +653,8 @@ function updateInvoice(invoice_id, updated_info, callback){
   let unit_charge = updated_info.unit_charge;
   let total_charge = updated_info.total_charge;
 
-  let sql = "UPDATE invoice SET ?? = ?, ?? = ?, ?? = ?, ?? = ?, ?? = ?, ?? = ?, ?? = ?, ?? = ?, ?? = ?, ?? = ? WHERE ?? = ?";
-  let inserts = ["tenant_id", tenant_id,
-                 "from_date", from_date,
-                 "to_date", to_date,
-                 "prior_read", prior_read,
-                 "cur_read", cur_read,
-                 "rubs", rubs,
-                 "has_submeter", has_submeter,
-                 "submeter_id", submeter_id,
-                 "unit_charge", unit_charge,
-                 "total_charge", total_charge,
-                 "invoice_id", invoice_id];
+  let sql = `UPDATE invoice SET tenant_id = ?, from_date = ?, to_date = ?, prior_read = ?, cur_read = ?, rubs = ?, has_submeter = ?, submeter_id = ?, unit_charge = ?, total_charge = ? WHERE invoice_id = ?`;
+  let inserts = [tenant_id, from_date, to_date, prior_read, cur_read, rubs, has_submeter, submeter_id, unit_charge, total_charge, invoice_id];
   connection.query(sql, inserts, function (err, result) {
     if (err) {
       console.log(`not able to update invoice info for invoice_id: ${invoice_id}  into database`);
@@ -726,8 +676,8 @@ function updateInvoice(invoice_id, updated_info, callback){
 // return true if delete successfully
 // return false if delete failed
 function deleteInvoice(invoice_id, callback){
-  let sql = `DELETE FROM invoice WHERE ?? = ?`;
-  let inserts = ["invoice_id", invoice_id];
+  let sql = `DELETE FROM invoice WHERE invoice_id = ?`;
+  let inserts = [invoice_id];
   connection.query(sql, inserts, function (err, result) {
     if (err) {
       console.log(err);
@@ -748,57 +698,111 @@ function deleteInvoice(invoice_id, callback){
 // associate meter with tenant
 // return true if add successfully
 // return false if add failed
-function associateMeterWithTenant(meter_id, tenant_id, callback){
-  let sql = `INSERT INTO meter_tenant(meter_id, tenant_id) VALUES(?,?)`;
-  let inserts = [meter_id, tenant_id];
-  connection.query(sql, inserts, function (err, result) {
-    if (err) {
-      console.log(`not able to add meter_id: ${meter_id} for tenant_id: ${tenant_id} into database`);
-      callback(false);
-    } else {
-      if (result.affectedRows == 1) {
-        console.log("added");
-        callback(true);
-      } else {
-        console.log(`not able to add meter_id: ${meter_id} for tenant_id: ${tenant_id} into database`);
-        callback(false);
-      }
-    }
-  });
-
-}
+// function associateMeterWithTenant(meter_id, tenant_id, callback){
+//   let sql = `INSERT INTO meter_tenant(meter_id, tenant_id) VALUES(?,?)`;
+//   let inserts = [meter_id, tenant_id];
+//   connection.query(sql, inserts, function (err, result) {
+//     if (err) {
+//       console.log(`not able to add meter_id: ${meter_id} for tenant_id: ${tenant_id} into database`);
+//       callback(false);
+//     } else {
+//       if (result.affectedRows == 1) {
+//         console.log("added");
+//         callback(true);
+//       } else {
+//         console.log(`not able to add meter_id: ${meter_id} for tenant_id: ${tenant_id} into database`);
+//         callback(false);
+//       }
+//     }
+//   });
+//
+// }
 
 // delete meter_tenant
 // return true if delete successfully
 // return false if delete failed
-function deleteMeterTenantRelation(meter_id, tenant_id, callback){
-  let sql = `DELETE FROM meter_tenant WHERE ?? = ? AND ?? = ?`;
-  let inserts = ["meter_id", meter_id,
-                 "tenant_id", tenant_id];
-  connection.query(sql, inserts, function (err, result) {
-    if (err) {
-      console.log(err);
-      console.log(`not able to delete meter_id: ${meter_id} for tenant_id: ${tenant_id} into database`);
-      callback(false);
-    } else {
-      if (result.affectedRows == 1) {
-        console.log("deleted!");
-        callback(true);
-      } else {
-        console.log(`not able to delete meter_id: ${meter_id} for tenant_id: ${tenant_id} into database`);
-        callback(false);
-      }
-    }
-  });
-}
+// function deleteMeterTenantRelation(meter_id, tenant_id, callback){
+//   let sql = `DELETE FROM meter_tenant WHERE meter_id = ? AND tenant_id = ?`;
+//   let inserts = [meter_id, tenant_id];
+//   connection.query(sql, inserts, function (err, result) {
+//     if (err) {
+//       console.log(err);
+//       console.log(`not able to delete meter_id: ${meter_id} for tenant_id: ${tenant_id} into database`);
+//       callback(false);
+//     } else {
+//       if (result.affectedRows == 1) {
+//         console.log("deleted!");
+//         callback(true);
+//       } else {
+//         console.log(`not able to delete meter_id: ${meter_id} for tenant_id: ${tenant_id} into database`);
+//         callback(false);
+//       }
+//     }
+//   });
+// }
 
-// get meter to tenant list(property_id)
 // this function will return a list of meter-tenant relations existed in given property_id
+// function selectMeterTenantListByProperty(property_id, callback){
+//   let sql = `
+//         select
+//             meter.meter_id,
+//             meter_tenant.tenant_id
+//         from
+//             meter inner join meter_tenant on meter.meter_id = meter_tenant.meter_id
+//         where meter.property_id = ?
+//         order by meter.meter_id`;
+//   let inserts = [property_id];
+//   connection.query(sql, inserts, function (err, meterTenantList) {
+//     if (err) {
+//       console.log(`not able to select meterTenantList of property_id: ${property_id} from database`);
+//       callback(false);
+//     } else {
+//       console.log(`property_id: ${property_id} meterTenantList list returned`);
+//       callback(meterTenantList);
+//     }
+//   });
+// }
 
+// this function will return a list of meter-submeter bill by property_id and given time period
+// function selectMeterSubmeterBillByProperty(property_id, from_date, to_date, callback){
+//   let sql = `
+//       with ?? as
+//         (
+//           select ?? from ?? where ?? = ?
+//         )
+//       select * from ??, ?? inner join ?? on ?? = ??
+//       where ?? = ?? and ?? >= ? and ?? <= ? `;
+//   let inserts = [
+//       "meter_list",
+//       "meter_id", "meter", "property_id", property_id,
+//       "meter_list", "bill", "submeter_bill", "bill.bill_id", "submeter_bill.bill_id",
+//       "meter_list.meter_id", "bill.meter_id", "bill.from_date", from_date, "bill.to_date", to_date];
+//
+//   connection.query(sql, inserts, function (err, meterSubmeterBillList) {
+//     if (err) {
+//       console.log(`not able to select meterSubmeterBillList of property_id: ${property_id} between ${from_date} and ${to_date} from database`);
+//       callback(false);
+//     } else {
+//       console.log(`property_id: ${property_id} between ${from_date} and ${to_date} meterSubmeterBillList returned`);
+//       callback(meterSubmeterBillList);
+//     }
+//   });
+//
+// }
 
-// get all meter and submeter bill(property_id, from_date, to_date)
-// get all meter and submeter (property_id)
-
+// this function will return a list of meter_id and submeter_id associate with given property
+// function selectAllMetersSubmetersByProperty(property_id, callback){
+//   let sql = `
+//       with ?? as(
+//         select meter_id from meter where property_id = 1
+//       )
+//       select submeter_id, meter_list.meter_id, tenant_id, multiplier
+//       from submeter,meter_list
+//       where submeter.meter_id = meter_list.meter_id`;
+//   let inserts = [
+//       "meter_list"
+//   ];
+// }
 
 // fetch bill by meter
 // fetch bill by account
@@ -826,8 +830,9 @@ exports.selectAllMeters = selectAllMeters;
 exports.insertMeter = insertMeter;
 exports.deleteMeter = deleteMeter;
 
-exports.associateMeterWithTenant = associateMeterWithTenant;
-exports.deleteMeterTenantRelation = deleteMeterTenantRelation;
+// exports.associateMeterWithTenant = associateMeterWithTenant;
+// exports.deleteMeterTenantRelation = deleteMeterTenantRelation;
+// exports.selectMeterTenantListByProperty = selectMeterTenantListByProperty;
 
 exports.selectAllSubmeters = selectAllSubmeters;
 exports.insertSubmeter = insertSubmeter;
@@ -846,4 +851,4 @@ exports.insertInvoice = insertInvoice;
 exports.updateInvoice = updateInvoice;
 exports.deleteInvoice = deleteInvoice;
 
-
+// exports.selectMeterSubmeterBillByProperty = selectMeterSubmeterBillByProperty;
